Require only one host to save the updated DGN Registry
# Description

When registering a standing query against a Streaming Graph cluster, one
host (the host which receives the API call) is meant to be responsible
for setting the persistor state. This delegation of responsibility is
maintained by the part of the code that writes the standing query and
the part of the code that writes the standing query outputs, but not the
part of the code that writes the updated DGN registry. This PR corrects
that oversight, and removes the default value which obscured it.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: c9f76a240533f68c4230cc34f0aa028891b439ad