Inline metric names for skimmability/legibility
# Description

There was an unnecessary level of code deduplication in the construction
of metric names. This made it very difficult to determine what a metric
would be called when quickly reading the code. This PR inlines all
metric name constructions, except for the `metricName` closure that
changes the name of some (but not all) metrics depending on which
product the user is running. This inconsistency was left in, but tagged
with a TODO, because changing it would require having our users update
their metrics dashboards (again).

## Type of change

- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: ee9cb355761d80d1b3ee371cad2f980bdc6913cf