Refactor/unify v1 v2 ingests in memory
# Support for v2 ingest objects in in-memory ingest map

This work is done in support of connecting version-2 ingests to API 2.

`QuineApp` and `QuineEnterpriseApp` store a reference in memory to
defined ingests
[here](https://github.com/thatdot/quine-plus/blob/97dcbf7f7f2c3ce289546da64f3ac389d1199a4b/public/quine/src/main/scala/com/thatdot/quine/app/QuineApp.scala#L134)
in a member `ingestStreams
: Map[NamespaceId, Map[IngestName,
IngestStreamWithControl[IngestStreamConfiguration]]]`. Most of the
handling of this map is addressed in
[IngestStreamState](https://github.com/thatdot/quine-plus/blob/97dcbf7f7f2c3ce289546da64f3ac389d1199a4b/public/quine/src/main/scala/com/thatdot/quine/app/routes/IngestRoutesImpl.scala#L28)
.

Since Ingest 2 objects
(`com.thatdot.quine.app.v2api.endpoints.V2IngestEndpointEntities.IngestConfiguration`)
are structured differently than the V1 ingest objects (i.e.
`IngestStreamConfiguration`) and this map is referenced in a fair number
of places (e.g. where viewing existing ingests in the api) this
collection becomes a problem for ingest2 objects. Creating a parallel
map creates a _lot_ of duplicate code, plus decisions about managing two
collections (e.g. Poll both collections to look for an ingest?)

To avoid this this PR replaces the `IngestStreamConfiguration` settings
variable defined inside `IngestStreamWithControl` with a union type
UnifiedIngestConfiguration = `Either[IngestConfiguration,
IngestStreamConfiguration]` so that V2 objects can be stored in the same
map without breaking anything.

Specifically, by being able to add ingest2 objects to this state, we
still should have
- starting and stopping ingests
- metrics/reporting access
- name checking for existence.
Note that this still doesn't address some discrepancies, such as storing
the ingest. For the moment, code that hasn't yet been extended to v2
ingests (e.g. persistor storage) is handled by filtering only on the
`Right[V1Ingest]` values.

There is conversion code included with the v2 objects so that they can
be viewed as V1 objects, so these collections can still work in quine
and streaming graph. Used with v1 objects only this PR should be a
no-op.

## Type of change

- [X] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [X] Unit test for round trip conversion v1 -> v2 -> v1 ingest for all
ingests.

# Checklist:

- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [ ] I have commented my code, particularly in hard-to-understand areas
- [X] I have added tests that prove my change is effective and works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any changes required of upstream modules have been updated and
have explicit tests in this PR

GitOrigin-RevId: 0fa59f33abee166fb7047d08bcb1738bf623dc09