Qu 2015 ingest 2 interface
# Some reworking of ingest2 to present a unified interface for extension

## Wrapping ingest2 in a simplified interface
This provides some simplification of the ingest 2 code while unifying
the interface. An example of usage can be seen
[here](https://github.com/thatdot/quine-plus/blob/fa4afbd479e3060e8ab90d83c6efce5f2e34954f/public/quine/src/test/scala/com/thatdot/quine/ingest2/sources/FramedSourceSpec.scala#L32)

### Changes:
- since CSV was the only edge case handled separately, manually parsing
this by frame using apache commons.csv. This is because the provided
pekko csv required special handling.
- There is a `FramedSourceProvider` labelling components that generate
framed sources. Not sure if this is at all necessary or even useful.
- Using a number of marker traits to label the providers with stream
features (e.g. compression). It might be better to apply these to the
framed sources directly.
- Fixed/simplified a number of sources.

- [X] Refactor

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [X] Added unit test to verify stream structure.

# Checklist:

- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have added tests that prove my change is effective and works
- [X] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 0837f6329f4656174b7de365c65e96774c2c817e