Only tell OpenCypher that we're doing a `YIELD *` when we are
# Description

We've been providing bad information to the openCypher AST (and its
semantic analysis steps) by telling it that all of our procedure calls
have a `YIELD *`. This led to nonsensical errors like "Query cannot end
with CALL combined with YIELD" on queries that had no `YIELD`. This
change uses the openCypher AST to determine whether the YIELD is a
`YIELD *`, then provides that information back in our
`QuineProcedureCall` AST node.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] New test added and passes

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 24f811567f01e75fa0ac2daeb4f2d8d5e9a56f55