fix for ingest statuses not updating properly
# Description

This fixes a bug where ingest statuses where not getting updated. That
is, calling the API `ingest pause` method would return a response that
displayed `pause`, but if you then retrieved the ingest again it would
display `running`. This surfaced in the end-to-end
`test_namespaced_ingest_pause_unpause` test (this tests bugs in the
enterprise namespaced version, but the bug could also be replicated by
creating, and then pausing an ingest in quine-oss, then retrieving the
ingest - the ingest would display "running" status)

The bug was due to changing the type of stored ingests in
`IngestStreamState` to an `Either[V1Ingest,V2Ingest]` type to deal with
V2ingests. As a part of that change, the types that where retrieved for
display were actually copies. Unfortunately the method that updates
ingest stream state calls a mutator (sets a var, actually) on the ingest
stream state. Since it was operating on a copy, the state would get
lost.

Changes:
- changed the `UnifiedIngestType` either to a case class, which
simplified some of the conversions.
- created an method that instead of retrieving ingests as v1
configurations (necessary for compatibility - returning a copied value)
returned references to the values stored in the ingest-stream state.

## Type of change

Delete options that are not relevant, add if necessary
- [X] Bug fix (non-breaking change which fixes an issue) expected)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [X] Tested with product-e2e-tests

# Checklist:

- [X] I have performed a self-review of my code
- [X] I have commented my code, particularly in hard-to-understand areas

GitOrigin-RevId: 6b907efee02fa65b0fcf256db1441351683c1ca6