Add support for modern SHOW PROCEDURES syntax via rewrite rule
# Description

`CALL dbms.procedures` is deprecated (`help.procedures` in our system).
We intend to continue supporting it, but for ease of discoverability,
this PR adds support for the modern GQL-style syntax equivalent, `SHOW
PROCEDURES`. We do this by essentially string-replacement, because the
overall structure of the `SHOW PROCEDURES` statement is so similar to a
regular procedure call anyways, it's just about a drop in replacement.
The one exception is a possible `EXECUTABLE` clause, which we simply
detect and ignore with a warning.

Also refactors the special-case statements (SHOW, EXPLAIN) to be more
extensible in the future.

The format `SHOW PROCEDURES WHERE <constraint>` is not supported and
will produce a syntax error similar to main.

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?

Ran and verified the output of the following queries
- [x] `SHOW PROCEDURES`
- [x] `SHOW PROCEDURES EXECUTABLE BY squidward`
- [x] `EXPLAIN MATCH (n) RETURN n`
- [x] `SHOW PROCEDURE EXECUTABLE YIELD name, description RETURN * LIMIT
2`
- [x] `SHOW PROCEDURE YIELD name WHERE name =~ 'cypher.*' RETURN *`
- [x] `CALL help.procedures()`

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: d185f1840e6f615f4e5e350e30a4a513aa4464c3