Make relayTell/relayAsk's direct-tell optimization threadsafe
# Description

While working on adding some wakeup protocol metrics, Christopher and I
identified a thread-safety issue in `relayTell` and `relayAsk`. Similar
to a few
years ago, we identified that the shard's nodes map is accessed from
off-thread. While the inner maps are concurrent and therefore
threadsafe, the outer map of namespaces is a `mutable.Map` and subject
to corruption if reads (from anywhere in the system performing a
`relayTell`/`relayAsk`) race writes (from the shard adding or removing a
namespace).

Simply switching to a concurrent.Map fixes this issue without notable
cost to performance, as concurrent.Map reads are non-blocking, and all
the writes were already happening on a single thread (the shard's).

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Documentation update

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 57b32d48622505a487f1489afff6f888eceb48ac