Add timers for node sleeps and wakes
# Description

Adds a pair of timers per shard-namespace:

One tracks how long it takes to put a node to sleep, from the time the
node sets its status to `ConsideringSleep` to the time the shard
completes its promise to remove the node from its caches.

The other tracks how long it takes to wake a node, from the time the
shard determines that a node it manages needs to be woken (because its
liveness status is `Nonexistent`) to the end of the node's constructor.

Also fixes a bug where sleep counters could be multiply-counted in some
error cases, and adds some documentation and line-shuffling to improve
legibility of the shard code.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Documentation update
- [x] New feature (non-breaking change which adds functionality)
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?

- [x] New metrics appear in default JMX reporter when running a recipe

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes

---------

Co-authored-by: Christopher Meade <2099307+cmeade@users.noreply.github.com>
GitOrigin-RevId: fba335cf8b678d63accc0a94f47fb9d226887b4c