Initial pass at error handling monad
# Description

This is an initial pass at an error handing concurrent monad.
Notable changes include:

- Added a FutureResult monad which takes an error type as well as a
result type. Under the hood, this just uses a future, but the interfaces
to this class return Either[E,T] rather than just a T (where E is the
error type and T is the result type), so it forces you to explicitly
handle errors of type E. Motivation for making Future a monad is
described below.
- Added a InterpM which is analogous to FutureResult but for Source
- Changed the base error type that all of our exceptions extend to
itself be a Throwable. This is so it can be used with the Future and
Source API with no overhead
- Changed the result of the interpreter to use InterpM[CypherException,
T] rather than Source[T, _]
- Changed the result of evaluating expressions that could previously
throw a CypherException to be an Either[CypherException, T] type. The
motivation for this is the same as for FutureResult (described below)
- Changed the way that exception serialization works. Unfortunately, now
we need to explicitly enumerate the errors we can handle in the pickler.
The alternative was to place all possible exceptions into one file and
make the base error a sealed trait
- Made BaseErrors not track stack traces by default (including all
subtypes like CypherException)

Notes:
For FutureResult and InterpM Since the underlying value is just a normal
future or source, any code that hasn't been updated to use
FutureResult/InterpM can just call `unsafeFuture` / `unsafeSource`,
which will provide the underlying future or source that will behave
exactly the same way as if we had been using a regular future the whole
time. For example, `InterpM.liftUnsafe(source).unsafeSource` is
functionally equivalent to just `source`

Also any exceptions thrown inside of a FutureResult or InterpM will
behave the same as previously, for example, the following code is
functionally equivalent

```
InterpM.error(CypherExecption.RunTime(""))
```
and
```
InterpM.liftUnsafe(Source.lazySingle( () =>
  throw CypherException.RunTime("")
))
```

# Motivation for creating FutureResults/InterpM/using Either

Something that has been talked about a lot is that we really shouldn't
be throwing exceptions for normal expected behavior, and returning an
error for invalid user input is normal expected behavior. Throwing
exceptions should be reserved for exceptional behavior (e.g. OOM, the
persister crashing, etc). So to this end, rather than throwing an
exception in functions where a CypherException is a valid possible
result, it should return a Either[CypherException, T] rather than just a
T, so we know to handle that case. FutureResult/InterpM were created to
be an equivalent interface for Futures and Soruces.

# Add on tasks

- Anywhere that is calling unsafeFuture, unsafeSource, evalUnsafe, or
getOrThrow should be updated to use the correct monad. There were dozens
of places that expected the old result from the interpreter or
expression evaluation that should be correctly updated.
- There are a few places in the interpreter where it is still using
evalUnsafe or getOrThrow
- We should do a deep dive into the code for anywhere an exception could
be raised and determine if using one of these Monads would be
appropriate in each case.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?

Repeated tests from https://github.com/thatdot/quine-plus/pull/3108

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
~- [ ] I have made corresponding changes to the documentation~
~[ ] I have added tests that prove my change is effective and works~
- [x] New and existing unit tests pass locally with my changes
~- [] Any changes required of upstream modules have been updated and
have explicit tests in this PR~
~- [] Any downstream modules depending on this change have an
integration plan described in the description~

GitOrigin-RevId: 0522c0b301a2540822d64fc3f7a2c96fe34abd8e