Add Ingest stream per-record timer metrics
# Description

Adds two metrics:

The first times the execution of each ingest query execution

The second times the deserialization of each ingested record (upstream
record, if there is a fan-out). Naturally, this only works on `Streamed`
(v1) / `Framed` (v2) sources, which have a native notion of what a
"record" even is.

Also, loosens the coupling of the `IngestMeter` utility, which is badly
in need of a refactor (unfortunately, the refactor took more than the 30
minutes I allotted to scoping it out and so will need to be done later).
Makes the metric name flag less product-sensitive and more descriptive

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?

- [x] New metrics appear in default JMX reporter

<img width="299" alt="image"
src="https://github.com/user-attachments/assets/90886615-845c-4dfe-bb98-0715af1878ab">

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 8a2af6405e6652f8c60c67770325b03b6821bbae