Qu 2070 refactor source formats
# Description

Some of our formats can only apply to file or streaming types. This
split existed in V1 ingests. Removing this division in V2 ingests made
some impossible combinations available. This restores this division. As
a side effect of this, the format type becomes a part of the source
definition, since the source needs to constrain the type of the
formatting.

Additional changes in this PR are mostly mechanical; since the structure
of the ingest specification changes a bit, these changes propagate
throughout the consumers of that type. In some places this enabled
removing of code paths for matching, since the illegal matches became
impossible.

GitOrigin-RevId: 26b80ebb736e2997ae6762a9eeae46f8d362a19f