Log on collisions in MVSQ part hashing
# Description

Collisions in ValueConstraint hashing (for MVSQ part IDs) cause issues
such as QU-1503.

For example, the following standing queries both hash to
`MultipleValuesStandingQueryPartId(22d4978b-c702-2aa9-65e5-1f4096812b7d)`
```diff
+ LocalProperty(Symbol(tag),Equal(Str(target)),None,Omitted)
- LocalProperty(Symbol(tag),NotEqual(Str(target)),None,Omitted)
```

This is because the MultipleValuesStandingQuery hasher, while totally
consistent in its hashes, is entirely based on structure. Two
structurally-equal queries, like the above, will have the same hashed
part ID. When this is detected, either while indexing an MVSQ or while
registering it on a node, we now warn or error to indicate the problem
so that it's just "wrong" instead of "wrong and confusing".

Also updated the scaladocs on cypher Expr hashing to explain why it's
safe

GitOrigin-RevId: e946548bf182eeaddbe15d85049c49e99755862c