OPTIONAL MATCH ergonomics: Ignore `null` values when querying for nodes/edges specifically
# Description

When using the node or edge query endpoints (or the associated RPCs via
websocket) the explicit expectation of the provided query is that it
returns only nodes/edges. However, this expectation is particularly easy
to accidentally violate with `OPTIONAL MATCH`, which creates
`NODE?`/`EDGE?`-type values of `null`. This PR adds handling to silently
ignore such values, while continuing to generate user-facing errors when
non-null values that don't match the asserted type are returned.

GitOrigin-RevId: 3da4a5510c4b9ba38116e2cc78592e7b87eeede4