Fix missing results from LocalProperty after node restart
# Description

The `LocalProperty` state doesn't persist any internal state, and it
used its internal state to serve `readResults`. The internal state would
be re-computed on the first call to `onNodeEvents`, but in some cases,
`onNodeEvents` wasn't called before `readResults`.

To fix this, we change `readResults` to work over the properties that
are provided to the `readResults` call rather than the internal state of
the `LocalPropertyState`. This is consistent with the behaviour of
`AllPropertiesState`, as well as the scaladoc on readResults itself. It
also makes `readResults` function correctly even if `onNodeEvents`
hasn't yet been called

Also updates some comments to clarify what state is vs isn't persisted

GitOrigin-RevId: 8d3c945d210c2be0aa9b6942248fd48914791b3d