Fix MVSQ false positives on SubscribeAcrossEdgeState
# Description

Fixes 2 bugs in SubscribeAcrossEdgeState and adds a regression test for
them

- When a node has a match based on an edge subscription, and another
edge exists with the same type and direction, and the original matched
edge is deleted, the SQ failed to invalidate the match. In other words,
we “forget to re-check” the remote side of edge subscriptions on
EdgeRemoved events in SubscribeAcrossEdge

- When a node has no edges that match a standing query, the node returns
an "I don't know" answer for that standing query, instead of a "I don't
match" answer.

GitOrigin-RevId: 25825e7bd6f85edeb6cce2aad8022b45a2956890