Refactor "Create MVSQ" message handler, simplify MVSQ initialization sequence
# Description

Simplifies and deduplicates code in the MVSQ creation message handler,
reducing the responsibilities of `onNodeEvents` and `onInitialize`
during initialization by unilaterally calling `readResults` to generate
the initial result group for a query.
Basically, `onNodeEvents` now does 2 things:
- On node wake / query registration, initialize the internal state of
the SQ state according to the replayed events passed in (_without_
emitting results)
- On thoroughgoing node events, react to those events to update the
internal state of the SQ state (_and_ emit results)

`readResult` is used to generate the first batch of results following a
query registration, as well as the results sent to each new subscriber
to a query. In other words, every subscriber now gets results via the
same means, whether they are the first or the hundredth.

Some additional checks were added to explicitly _prevent_ `onNodeEvents`
from emitting matches during its first run, but this is an optimization
to reduce intermediate results (and message-sending), and is documented
as such.

GitOrigin-RevId: 9a54a9d519178f66620491ac4daca3d7f20e92b7