QU-1888: Update quine-query-language dependency, update tests
# Description

This PR bumps the version of `quine-query-language` from `0.1.14` to
`0.1.16`. The bump from `0.1.14` -> `0.1.15` represented a super small
change in this PR, which I've commented on. The bump from `0.1.15` ->
`0.1.16` represents the changes to the language server I made.

The language server dependency now explicitly lists its capabilities via
the following providers sent after the `initialize` request:
- `semanticTokensProvider `
- `completionProvider `
- `diagnosticProvider `

To assert these changes, I've updated the tests to assert the
capabilities returned by our language server. While I was in those
tests, I also switched off of using hard-coded json strings, to using
ADT's with circe, to make it a bit easier to write and follow.

Again, aside from some very minor changes for the bump to `0.1.15`, the
entirety of this PR is the bump of the dependency, and updates to the
tests.

GitOrigin-RevId: d9fb0c6d90682e4fd585a86a582d3573eb438a61