QU-2114: Extend telemetry data to include recipe title and contents hash
# Description

If a recipe is used, ImproveQuine telemetry is extended to include a
secure hash of the recipe's title and material contents. This feature is
intended to be able to detect iteration of recipes.

"material contents" refers to syntactically relevant fields to the Quine
graph. It excludes `contributor`, `summary`, `description`, and
`iconImage` fields as well as whitespace/formatting of the actual recipe
yaml file. The implementation only hashes the relevant fields of the
parsed object in Quine.

GitOrigin-RevId: 43e381bf1dbda739cbdcd311f318f3476cb76ae4