Qu 2082/omit metadata from sq output
# Description

Added an option to standing query outputs to include the metadata in the
standing query output or not (for both v1 and v2 api). If the metadata
is set to be included the output will look like

```
{
 "meta":the_metadata,
 "data": the_sq_outout
}
```

while if it is set to be excluded then the result will just be
`the_sq_output` not contained within an extra object.

GitOrigin-RevId: 5a21a898dd2d7974401d7aaf4015222d6c0ab0e6