Fix Quine logo when rendered on Safari on latest macOS update
# Description

The latest update of macOS renders the Quine logo across the entire
width of the `TopBar` component in the header. This PR makes that image
a child of a Flex child container by wrapping it in a div, ensuring it
is positioned correctly, and ensures that the image has a minimum width,
so that it doesn't collapse to 0 width.

# Images

## Before

![image](https://github.com/user-attachments/assets/2e49321e-ff93-4dc8-861b-02580f4e6e73)

## After

![image](https://github.com/user-attachments/assets/28a1fcf0-2177-4ab7-8aff-23bfd221ad04)

GitOrigin-RevId: 826fe159eb80a251ff6ce1c4f89a4da7214425fb