Remove use of `SLECT COUNT(*)`
COUNT isn't supported on AWS Keyspaces.

Also, the count is just being checked for `> 0`. Counting all items in
the table is an inefficient way to check if the table's empty. There's
already an existing method for checking if the table's empty, based on
seeing if `SELECT quine_id LIMIT 1` returns any results.