Make Transformations Consistent With V2 Creation/List Patterns
Following #3635, this makes Transformation endpoints in Novelty follow
the creation and list patterns of other resources. Additionally, this
allows Novelty Ingest Streams to be created with either a transformation
name or body, which sets us up for API evolution toward the referential
transparency with versioning model we have in mind.

GitOrigin-RevId: 8b6f28b131d1de3218221d44c47f9407d1be193d