Fix input2 JSON parsing test to use new format
Our API V2 documentation said the JSON format we had worked on JSON
lines files. Underneath, it actually worked with whatever the pekko
connector supported. This is a problem, because JSON lines allows top
level values that are any JSON value, not just objects, while the pekko
framing is more flexible on the delimiter, but requires object elements.
Now there are two formats to support these different varieties of
flexibility, and the test now uses the JSON (not lines) format for the
test that needs delimiter flexibility.

GitOrigin-RevId: e948c76d4f878802f41e2a2f1a43ed34df6b0bd9