rename SecureRandom to StrongUUID
rename SecureRandom to StrongUUID so that the type better describes what
it does.

# Description

SecureRandom was not named well. Its intention is to be an alternative
to 'UUID.randomUUID', and 'SecureRandom.randomUUID' is less meaningful
as a replacement than the new 'StrongUUID.randomUUID'. Now, as a
developer you can use java's UUID type, or our StrongUUID type in
situations where FIPS 140 cryptography is required.

GitOrigin-RevId: 6ed699a98b7fe633cec4082b365ec6fd9b5a5289