Increase max SSE line and event sizes to 5MB
# Description

Increases the default maximum line and event size for SSE messages to 5
MB. Previously, if a message exceeded the earlier limit (which was 4k)
Quine would fail to parse the message and retry at the same point in the
stream infinitely. This made it look like Quine was getting stuck at
that point in the stream. New [changes have been merged into
Pekko](https://github.com/apache/pekko-http/pull/744) to address [the
underlying problem in the Pekko
connector](https://github.com/apache/pekko-connectors/pull/1205), but a
larger default line and event size is also a positive change, and will
minimize the impact until the next version of Pekko is released.

GitOrigin-RevId: c63829b4ecad749ecb39ee720b6d58d2a13439fd