Make Query UI faster.
This adds a new Cypher Procedure GetFilteredEdges which will return the edges needed by the UI but without having to visit every node connected to the node in question. Instead it will filter edges on the node in question first, and only visit the nodes that need to be visited. This can have a large effect if some of the needlessly visited nodes were supernodes or not in memory. This will be incorporated into the query optimization phase at a later point.

GitOrigin-RevId: e457b44a8e496ace15f0d743355e52d6c36dd97b