Fixing some failing tests that were enabled on the PR this is merging into
This PR turns back on all tests in CI and should enable them to be ran locally quicker and more reliably.  There is now a separate test tag `sbt licenseTest:test` for tests that require a license server that will be excluded for sbt test and some broken tests were fixed.

GitOrigin-RevId: 576b34706d2f66e06edcb6c71b03fb4e9dad2fd7