Filter uninteresting nodes
# Description

When performing an all node scan query, we shouldn't return nodes that
do not have an interesting present (no edges, properties, or labels) at
the atTime of the query. While these nodes may have an interesting
history (they used to have edges, properties, or labels), if the don't
have an interesting present, then they should be excluded. This behavior
would be more intuitive to users who are confused when nodes that are
DETACH DELETEd show in a `match (n) return n` query.

GitOrigin-RevId: ebbc27d3957a0ee5c96bbc239a3406b780b54a4d